resolve errors and update Maintenance page UI
